==> tutos/04_fonctions_boucles.qmd <==
---
title: "Fonctions et boucles"
author: "A. Peris, Avignon Université"
date: "01/16/2025"
format: html
---

Packages

```{r}
#| message: false
#| warning: false
library(sf)
library(ggplot2)
library(dplyr)
library(ggspatial)
```

On commence par charger les données sur le périmètre des communes. 

```{r, results='hide'}
#| echo: true
#| message: false
#| warning: false
com <- st_read("../data/COMMUNE.shp")
```

## Les fonctions

### Définition et cas d'utilisation

Une **fonction** est un outil qui permet d’exécuter une série d’instructions en prenant des entrées (appelées arguments) et en produisant une sortie (le résultat). L'exemple ci-dessous, qui reponse sur la création d'une fonction additionnant deux arguments illustre la syntaxe de base d'une fonction en R.

```{r}
# Définition d'une fonction
nom_de_la_fonction <- function(argument1, argument2) {
  # Instructions à exécuter
  resultat <- argument1 + argument2 # Exemple : addition des arguments
  
  # Retourner le résultat
  return(resultat)
}

# Appel de la fonction
nom_de_la_fonction(3, 7) # Résultat : 10
```

### Application

On peut enchainer toute une série d'instruction dans une fonction. 

```{r}
#| message: false
#| warning: false
plot_dpt <- function(code_departement){
  
  comdep <- com %>% 
    filter(INSEE_DEP == code_departement)
  
  comctr <- st_point_on_surface(comdep)
  
  troisplus <- comctr %>% 
    arrange(desc(POPULATION)) %>% 
    slice(1:3)
  
  g <- ggplot()+
    geom_sf(data = comdep, fill="white")+
    geom_sf(data = comctr, aes(size=POPULATION), alpha=.6)+
    geom_sf_text(data = troisplus, aes(label=NOM), nudge_y = -3000)+
    scale_size_area(max_size = 5)+
    annotation_scale()+
    theme_bw()+
    labs(title = paste("La population communale dans le", code_departement))

  return(g)  
}

plot_dpt("47")
```

## Les boucles

### Définition et cas d'utilisation

Une boucle est une structure qui permet de répéter automatiquement une série d'instructions tant qu'une condition est remplie ou pour chaque élément d'une séquence.

Les boucles sont utilisées pour :

- Automatiser des tâches répétitives : Exécuter le même code plusieurs fois sans le réécrire. 
- Parcourir des données : Itérer sur les éléments d'un vecteur, d'une liste ou d'une table.
- Gérer des conditions dynamiques : Répéter une action jusqu’à ce qu’une certaine condition soit remplie.

Il existe plusieurs types de boucles. Les boucles `for` servent à effectuer des itérations à partir d'un vecteur.

```{r}
# Boucle pour afficher les chiffres de 1 à 5
for (i in 1:5) {
  print(i) # Affiche la valeur de i à chaque itération
}
```

Les boucles `while` permettent de répéter une action tant qu'une condition est vraie.

```{r}
# Boucle pour afficher les chiffres de 1 à 5
x <- 1
while (x <= 5) {
  print(x) # Affiche la valeur de x
  x <- x + 1 # Incrémente x pour éviter une boucle infinie
}
```

### Application 

On peut se servir des boucles pour automatiser des tâches. Dans l'exemple ci-dessous, on modifie légèrement la fonction pour faire une carte de population ajoutant une instruction d'export, puis on fait un export automatique des cartes de toute la région Sud-PACA. 

```{r results='hide'}
#| message: false
#| warning: false
print_dpt <- function(code_departement){
  
  comdep <- com %>% 
    filter(INSEE_DEP == code_departement)
  
  comctr <- st_point_on_surface(comdep)
  
  troisplus <- comctr %>% 
    arrange(desc(POPULATION)) %>% 
    slice(1:3)
  
  g <- ggplot()+
    geom_sf(data = comdep, fill="lightgrey", color="white")+
    geom_sf(data = comctr, aes(size=POPULATION), alpha=.6)+
    geom_sf_text(data = troisplus, aes(label=NOM), nudge_y = -3000)+
    scale_size_area(max_size = 5)+
    annotation_scale()+
    theme_bw()+
    labs(title = paste("La population communale dans le", code_departement))

  path <- paste0("../outputs/carto_dep/", "cartopop", code_departement, ".png")
  
  ggsave(path, g, dpi = 300)  
}


for(i in c("06", "04", "05", "13", "83", "84")){
  print_dpt(i)
}
```
On peut maintenant se rendre dans le dossier `/outputs/carto_dep/` où ont été sauvegardées toutes les cartes de population des départements de la région Sud-PACA.
